[Add] use `xxx.registrant.ts` form for any registrations related code
