+++ src/code/electron/registrant.ts
@@ -1,0 +1,1 @@
+import 'src/code/platform/configuration/electron/configuration.registrant';
