+++ src/code/platform/configuration/browser/configuration.registrant.ts
@@ -1,0 +1,29 @@
+import { BuiltInConfigScope, IConfigRegistrant } from "src/code/platform/configuration/common/configRegistrant";
+import { DefaultConfigStorage } from "src/code/platform/configuration/common/configStorage";
+import { Registrants } from "src/code/platform/registrant/common/registrant";
+
+class DefaultUserConfiguration extends DefaultConfigStorage {
+    protected createDefaultModel(): Record<PropertyKey, any> {
+        return {
+            'workbench': {
+                language: 'en',
+                keyboardScreenCast: false,
+            },
+            // old
+            'notebookManager': {
+                notebookManagerExclude: [
+                    '^\\..*',
+                ] as string[],
+                notebookManagerInclude: [
+
+                ] as string[],
+                previousOpenedNotebook: '',
+            },
+        };
+    }
+}
+
+(function registerBrowserDefaultConfiguration() {
+    const Registrant = Registrants.get(IConfigRegistrant);
+    Registrant.registerDefaultBuiltIn(BuiltInConfigScope.User, new DefaultUserConfiguration());
+})();
