+++ src/code/platform/configuration/common/configuration.ts
@@ -197,41 +197,3 @@ export class DefaultUserConfigModel extends __ConfigModel {
     }
 
 }
-
-class DefaultApplicationConfiguration extends DefaultConfigStorage {
-    protected override createDefaultModel(): Record<PropertyKey, any> {
-        return {
-            'application': {
-                displayLanguage: 'en',
-                keyboardScreenCast: false,
-            },
-            'notebookManager': {
-                defaultConfigOn: false,
-                startPreviousNotebookManagerDir: true,
-                previousNotebookManagerDir: '',
-            },
-        };
-    }
-}
-
-class DefaultUserConfiguration extends DefaultConfigStorage {
-    protected createDefaultModel(): Record<PropertyKey, any> {
-        return {
-            'notebookManager': {
-                notebookManagerExclude: [
-                    '^\\..*',
-                ] as string[],
-                notebookManagerInclude: [
-
-                ] as string[],
-                previousOpenedNotebook: '',
-            },
-        };
-    }
-}
-
-(function registerDefaultConfiguration() {
-    const Registrant = Registrants.get(IConfigRegistrant);
-    Registrant.registerDefaultBuiltIn(BuiltInConfigScope.Application, new DefaultApplicationConfiguration());
-    Registrant.registerDefaultBuiltIn(BuiltInConfigScope.User, new DefaultUserConfiguration());
-})();
