--- src/code/platform/configuration/electron/configuration.register.ts
@@ -1,4 +1,0 @@
-import { BuiltInConfigScope, IConfigRegistrant } from "src/code/platform/configuration/common/configRegistrant";
-import { DefaultConfigStorage } from "src/code/platform/configuration/common/configStorage";
-import { Registrants } from "src/code/platform/registrant/common/registrant";
-
